feat: extract global component check to custom hook
